Decouple uart backend for GVSoC and RTL Simulations
