+++ generic_runtime/inc/uart_setup.h
@@ -1,0 +1,16 @@
+// Copyright 2025 ETH Zurich and University of Bologna.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+//
+// Victor Jung <jungvi@iis.ee.ethz.ch>
+
+#ifndef UART_SETUP_H
+#define UART_SETUP_H
+
+#include <stdio.h>
+
+void _uart_init(void);
+void _uart_deinit(void);
+char uart_putc(char c, FILE *file);
+
+#endif // UART_SETUP_H
