+++ generic_runtime/src/alloc.c
