+++ generic_runtime/src/uart_deinit_gvsoc.c
@@ -1,0 +1,9 @@
+// Copyright 2025 ETH Zurich and University of Bologna.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+//
+// Victor Jung <jungvi@iis.ee.ethz.ch>
+
+#include "uart_setup.h"
+
+void _uart_deinit(void){ }
